refactor(hashtower): go back to the class and one-function-one-file style
